+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Behavior:
 
  * `apictl traffic initialize` no longer waits for the traffic-proxy to become ready before exiting.
  * Feature: Developer portal will show swagger documentation for up to five services (or more with appropriate license)
+ * Feature: local-devportal is now a standalone go binary with no external dependencies
 
 Other: